+++ src/main/java/mod/gottsch/forge/treasure2/core/capability/PouchCapability.java
@@ -22,10 +22,9 @@ import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ForgeCapabilities;
 import net.minecraftforge.common.capabilities.ICapabilitySerializable;
-import net.minecraftforge.common.capabilities.RegisterCapabilitiesEvent;
 import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.items.CapabilityItemHandler;
 import net.minecraftforge.items.ItemStackHandler;
 
 /**
@@ -42,7 +41,7 @@ public class PouchCapability implements ICapabilitySerializable<CompoundTag> {
 	@Override
 	public <T> LazyOptional<T> getCapability(Capability<T> cap, Direction side) {
 
-		if (cap ==CapabilityItemHandler.ITEM_HANDLER_CAPABILITY) {
+		if (cap ==ForgeCapabilities.ITEM_HANDLER) {
 			return LazyOptional.of(() -> itemHandler).cast();
 		}
 		return LazyOptional.empty();